GUI
Window with bunch of buttons, that can do nothing, except react on the
clicking.  Download the hole archive, called Interface, open try3.exe,
in order to see the window. If you are interested in source code files,
just open try.pro (you can find it in the folder source, inside the
Interface archive), to open it you have to own Qt5