New Main Window
I deleted the previous one. This version can do lots of stuff. It can
scan directory, it can close windows by itself.  It can't hide the
buttons, but I have no idea why. It always getting broken, when you try
to scan the root, but it will be developed later