1.0.0
This version uses md5 signature scanner to find viruses.
Was tested on real malware, antivirus detects it quickly if has the
needed signature in the base. Its about 130 000 signatures as far as i
remember.
When virus is detected, you can delete it, quarantine (file becomes
encrypted and cannot do any bad things), or ignore it. The program also
comes with dqr.exe, that takes the encrypted files out of quarantine.
Next release is coming in the next 14 hours (i'll wake up and make a
release with installer and installing/using guide)